[Templates] Better NetStandard template
