--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.NetStandard/.template.config/template.json
@@ -1,20 +1,0 @@
-{
-    "author": "MonoGameTeam",
-    "classifications": [ "MonoGame" ],
-    "name": "MonoGame NetStandard Library",
-    "groupIdentity": "MonoGame.BlankApp.NetStandard",
-	"identity": "MonoGame.BlankApp.NetStandard.CSharp",
-    "shortName": "mgnetstandardlib",
-    "tags": {
-        "language": "C#",
-        "type": "project"
-    },
-    "primaryOutputs": [ 
-                        {
-                            "path": "BlankApp.csproj"
-                        },
-                      ], 
-    "sourceName": "BlankApp", 
-    "preferNameDirectory" : "true",
-    "description": "A MonoGame NetStandard project. Can be used to share game code between platforms."
-}
