+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.NetStandard/.template.config/dotnetcli.host.json
