+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.NetStandard/.template.config/template.json
@@ -1,0 +1,22 @@
+{
+    "author": "MonoGame Team",
+    "classifications": [
+        "MonoGame"
+    ],
+    "name": "MonoGame NetStandard Library",
+    "groupIdentity": "MonoGame.Library.NetStandard",
+    "identity": "MonoGame.Library.NetStandard.CSharp",
+    "shortName": "mgnetstandard",
+    "tags": {
+        "language": "C#",
+        "type": "project"
+    },
+    "primaryOutputs": [
+        {
+            "path": "MgNamespace.csproj"
+        }
+    ],
+    "sourceName": "MgNamespace",
+    "preferNameDirectory": "true",
+    "description": "A MonoGame NetStandard library project. Can be used to share game code between platforms."
+}
