+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.NetStandard/MGNamespace.csproj
@@ -5,7 +5,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-     <PackageReference Include="MonoGame.Framework.Portable" Version="3.6.0.1625" />
+     <PackageReference Include="MonoGame.Framework.DesktopGL.Core" Version="3.7.0.7">
+      <PrivateAssets>All</PrivateAssets>
+    </PackageReference>
   </ItemGroup>
 
 </Project>
